server/requirements.txt
Flask==2.3.2
Flask_Cors==4.0.0
itemadapter==0.8.0
psycopg2-binary==2.9.6
Scrapy==2.9.0
Flask-WTF==1.1.1